refactor: [SRTP-419] mode WebClient.builder bean creation into ApplicationConfig class
